requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
numpy==1.19.0
Pillow==7.2.0
requests==2.24.0
RPi.GPIO==0.7.0
soupsieve==2.0.1
spidev==3.5
urllib3==1.25.10
wikipedia==1.4.0
